solver: split llbsolver pkg
